Simmods, and programs have been added to comms, decks, and rccs
I removed the mods object in all the devices, because it would mirror
programs for all devices, so I removed it and replaced it with
programlist, but that doesn’t get made until the program makes the
device in the intentory.

Added the all programs to buy. Added code to support agent prices up to
rating 6, but since you can only buy them at rating 4 at chargen, a lot
of the extra effert and code is unnecessary.  But I’ll leave it in for
now.

Commlinks can now have sim mods and hotsim. I guess hotsim is suppose
to raise the avail of commlinks…but I don’t really care…maybe I’ll come
back and fix that later.